move `shell_call` into main library
