+++ examples/nips17_adversarial_competition/eval_infra/code/worker.py
@@ -39,6 +39,7 @@ import uuid
 import eval_lib
 
 from six import iteritems
+from cleverhans.utils import shell_call
 
 
 # Sleep time while waiting for next available piece of work
@@ -94,39 +95,6 @@ METADATA_JSON_TYPE_TO_TYPE = {
 CMD_VARIABLE_RE = re.compile('^\\$\\{(\\w+)\\}$')
 
 
-def shell_call(command, **kwargs):
-  """Calls shell command with argument substitution.
-
-  Args:
-    command: command represented as a list. Each element of the list is one
-      token of the comman. For example "cp a b" becomes ['cp', 'a', 'b']
-      If any element of the list looks like '${NAME}' then it will be replaced
-      by value from **kwargs with key 'NAME'.
-    **kwargs: dictionary with argument substitution
-
-  Returns:
-    output of the command
-
-  Raises:
-    subprocess.CalledProcessError if command return value is not zero
-
-  This function is useful when you need to do variable substitution prior
-  running the command. Below are few examples of how it works:
-
-    shell_call(['cp', 'a', 'b'], a='asd') calls command 'cp a b'
-
-    shell_call(['cp', '${a}', 'b'], a='asd') calls command 'cp asd b',
-    '${a}; was replaced with 'asd' before calling the command
-  """
-  command = list(command)
-  for i in range(len(command)):
-    m = CMD_VARIABLE_RE.match(command[i])
-    if m:
-      var_id = m.group(1)
-      if var_id in kwargs:
-        command[i] = kwargs[var_id]
-  logging.debug('Executing shell command: %s', ' '.join(command))
-  return subprocess.check_output(command)
 
 
 def make_directory_writable(dirname):
